потерял аннотации сервисов. все верно обвяил в бинах. разделил логику игры и юзеров
